# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Sammium NexusOps API
  version: 0.1.0
  description: API contract for the event-driven integration portfolio MVP.
servers:
  - url: http://localhost:4000
security:
  - ApiKeyAuth: []
paths:
  /health:
    get:
      security: []
      summary: Check PostgreSQL and Redis health
      responses:
        '200': { description: Healthy }
        '503': { description: Degraded }
  /api/v1/demo/payments:
    post:
      summary: Generate a Stripe-shaped demo payment event
      parameters:
        - $ref: '#/components/parameters/TenantId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                amount: { type: number, example: 4999 }
                currency: { type: string, example: PHP }
                customerName: { type: string, example: Aoki Developer }
                customerEmail: { type: string, format: email, example: aoki@example.com }
      responses:
        '202': { description: Event accepted and queued }
  /api/v1/events:
    get:
      summary: List integration events
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - { name: status, in: query, schema: { type: string } }
        - { name: provider, in: query, schema: { type: string } }
      responses:
        '200': { description: Paginated event collection }
  /api/v1/events/{eventId}:
    get:
      summary: Get an event and its action executions
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - { name: eventId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Event details }
        '404': { description: Not found }
  /api/v1/events/{eventId}/replay:
    post:
      summary: Requeue a failed or completed event
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - { name: eventId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '202': { description: Event requeued }
  /api/v1/workflows:
    get:
      summary: List workflow definitions
      parameters:
        - $ref: '#/components/parameters/TenantId'
      responses:
        '200': { description: Workflow collection }
    post:
      summary: Create a workflow
      parameters:
        - $ref: '#/components/parameters/TenantId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Workflow'
      responses:
        '201': { description: Workflow created }
  /webhooks/stripe:
    post:
      security: []
      summary: Receive and verify Stripe webhook events
      responses:
        '200': { description: Webhook received }
        '400': { description: Invalid signature or payload }
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
  parameters:
    TenantId:
      name: x-tenant-id
      in: header
      required: true
      schema: { type: string, format: uuid }
  schemas:
    Workflow:
      type: object
      required: [name, triggerType, actions]
      properties:
        name: { type: string }
        triggerType: { type: string, example: payment.succeeded }
        enabled: { type: boolean, default: true }
        conditions:
          type: array
          items:
            type: object
            properties:
              field: { type: string, example: data.amount }
              operator: { type: string, enum: [equals, notEquals, greaterThan, greaterThanOrEqual, lessThan, contains, exists] }
              value: {}
        actions:
          type: array
          items:
            type: object
            properties:
              type: { type: string, enum: [hubspot.upsertContact, slack.sendMessage, audit.record] }
              config: { type: object, additionalProperties: true }